requirements.txt
openai
newsapi-python
python-dotenv
cartesia
requests